refactor: split 2400-line App.tsx into modular architecture
  Extract 15+ modules for better separation of concerns, reducing main file by 77% while maintaining all
   functionality
