+++ src/renderer/constants/providers.ts
@@ -1,0 +1,23 @@
+/**
+ * Provider constants for terminal/agent integrations
+ * Lists all supported coding agent providers
+ */
+
+export const TERMINAL_PROVIDER_IDS = [
+  'qwen',
+  'codex',
+  'claude',
+  'droid',
+  'gemini',
+  'cursor',
+  'copilot',
+  'amp',
+  'opencode',
+  'charm',
+  'auggie',
+  'kimi',
+  'kiro',
+  'rovo',
+] as const;
+
+export type TerminalProviderId = typeof TERMINAL_PROVIDER_IDS[number];
